# SIMD: Optimize the performace of np.packbits in ARM-based machine.