--- a/numpy/core/src/multiarray/compiled_base.c
+++ b/numpy/core/src/multiarray/compiled_base.c
@@ -1505,12 +1505,14 @@ arr_add_docstring(PyObject *NPY_UNUSED(dummy), PyObject *args)
 #endif
 
 #if defined __ARM_NEON || defined __ARM_NEON__
-    #include<arm_neon.h>
-    int32_t _mm_movemask_epi8_neon(uint64x2_t input)
+    #include <arm_neon.h>
+    int _mm_movemask_epi8_neon(uint64x2_t input)
     {
-        int64x1_t m0 = vdup_n_s64(0);
-        uint64x2_t v0 = vshlq_u64(vshrq_n_u64(input, 63), vcombine_s64(m0, m0));
-        return (int)vgetq_lane_u64(v0, 0) + ((int)vgetq_lane_u64(v0, 1) << 1);
+        uint8x16_t ninput = vreinterpretq_u8_u64(input);
+        int8x8_t m0 = vcreate_s8(0x0706050403020100ULL);
+        uint8x16_t v0 = vshlq_u8(vshrq_n_u8(ninput, 7), vcombine_s8(m0, m0));
+        uint64x2_t v1 = vpaddlq_u32(vpaddlq_u16(vpaddlq_u8(v0)));
+        return (int)vgetq_lane_u64(v1, 0) + ((int)vgetq_lane_u64(v1, 1) << 8);
     }
 #endif
 /*
@@ -1552,6 +1554,7 @@ pack_inner(const char *inptr,
                 a = npy_bswap8(a);
                 b = npy_bswap8(b);
             }
+            
             /* note x86 can load unaligned */
             __m128i v = _mm_set_epi64(_m_from_int64(b), _m_from_int64(a));
             /* false -> 0x00 and true -> 0xFF (there is no cmpneq) */
@@ -1569,7 +1572,6 @@ pack_inner(const char *inptr,
     }
 #elif defined __ARM_NEON || defined __ARM_NEON__
     if (in_stride == 1 && element_size == 1 && n_out > 2) {
-        uint64x2_t zero = vdupq_n_u64(0);
         /* don't handle non-full 8-byte remainder */
         npy_intp vn_out = n_out - (remain ? 1 : 0);
         vn_out -= (vn_out & 1);
@@ -1583,11 +1585,10 @@ pack_inner(const char *inptr,
             }
             const npy_uint64 __attribute__((aligned(16))) data[2] = {a, b};
             uint64x2_t v = vld1q_u64(data);
+            uint64x2_t zero = vdupq_n_u64(0);
             /* false -> 0x00 and true -> 0xFF (there is no cmpneq) */
-            v = vreinterpretq_u64_u8(vceqq_u8(vreinterpretq_u8_u64(v),
-                                              vreinterpretq_u8_u64(zero)));
-            v = vreinterpretq_u64_u8(vceqq_u8(vreinterpretq_u8_u64(v),
-                                              vreinterpretq_u8_u64(zero)));
+            v = vreinterpretq_u64_u8(vceqq_u8(vreinterpretq_u8_u64(v), vreinterpretq_u8_u64(zero)));
+            v = vreinterpretq_u64_u8(vceqq_u8(vreinterpretq_u8_u64(v), vreinterpretq_u8_u64(zero)));
             /* extract msb of 16 bytes and pack it into 16 bit */
             r = _mm_movemask_epi8_neon(v);
             /* store result */