--- a/numpy/core/src/multiarray/compiled_base.c
+++ b/numpy/core/src/multiarray/compiled_base.c
@@ -1504,13 +1504,13 @@ arr_add_docstring(PyObject *NPY_UNUSED(dummy), PyObject *args)
 #include <emmintrin.h>
 #endif
 
-#if defined __ARM_NEON || defined __ARM_NEON__
-    #include <arm_neon.h>
-    int sign_mask(uint64x2_t input)
+#ifdef NPY_HAVE_NEON
+    typedef npy_uint64 uint64_unaligned __attribute__((aligned(16)));
+    static NPY_INLINE int32_t
+    sign_mask(uint8x16_t input)
     {
-        uint8x16_t ninput = vreinterpretq_u8_u64(input);
         int8x8_t m0 = vcreate_s8(0x0706050403020100ULL);
-        uint8x16_t v0 = vshlq_u8(vshrq_n_u8(ninput, 7), vcombine_s8(m0, m0));
+        uint8x16_t v0 = vshlq_u8(vshrq_n_u8(input, 7), vcombine_s8(m0, m0));
         uint64x2_t v1 = vpaddlq_u32(vpaddlq_u16(vpaddlq_u8(v0)));
         return (int)vgetq_lane_u64(v1, 0) + ((int)vgetq_lane_u64(v1, 1) << 8);
     }
@@ -1570,27 +1570,27 @@ pack_inner(const char *inptr,
             inptr += 16;
         }
     }
-#elif defined __ARM_NEON || defined __ARM_NEON__
+#elif defined NPY_HAVE_NEON
     if (in_stride == 1 && element_size == 1 && n_out > 2) {
         /* don't handle non-full 8-byte remainder */
         npy_intp vn_out = n_out - (remain ? 1 : 0);
         vn_out -= (vn_out & 1);
         for (index = 0; index < vn_out; index += 2) {
             unsigned int r;
-            npy_uint64 a = *(npy_uint64*)inptr;
-            npy_uint64 b = *(npy_uint64*)(inptr + 8);
+            npy_uint64 a = *((uint64_unaligned*)inptr);
+            npy_uint64 b = *((uint64_unaligned*)(inptr + 8));
             if (order == 'b') {
                 a = npy_bswap8(a);
                 b = npy_bswap8(b);
             }
-            const npy_uint64 __attribute__((aligned(16))) data[2] = {a, b};
-            uint64x2_t v = vld1q_u64(data);
+            uint64x2_t v = vcombine_u64(vcreate_u64(a), vcreate_u64(b));
             uint64x2_t zero = vdupq_n_u64(0);
             /* false -> 0x00 and true -> 0xFF (there is no cmpneq) */
             v = vreinterpretq_u64_u8(vceqq_u8(vreinterpretq_u8_u64(v), vreinterpretq_u8_u64(zero)));
             v = vreinterpretq_u64_u8(vceqq_u8(vreinterpretq_u8_u64(v), vreinterpretq_u8_u64(zero)));
             /* extract msb of 16 bytes and pack it into 16 bit */
-            r = sign_mask(v);
+            uint8x16_t input = vreinterpretq_u8_u64(v);
+            r = sign_mask(input);
             /* store result */
             memcpy(outptr, &r, 1);
             outptr += out_stride;