### BackGround:
np.packbits is widely used in image-processing and AI-model processing, However, the performance in ARM is bad than other platforms. As a result, SIMD optimization resulted in a performance improvement of 70~79%.
### System info:

- HardWare: KunPeng
- Processor: ARMv8 2.6GMHZ 8 processors
- OS: Linux ecs-9d50 4.19.36-vhulk1905.1.0.h276.eulerosv2r8.aarch64
### The optimization effect:
```
       before           after         ratio
     [1b212bda]       [f3b9c48b]
     <master>         <pack_bit_neon>
     22.2±0.02μs      6.65±0.03μs     0.30  bench_core.PackBits.time_packbits(<class 'bool'>)
        395±1μs       82.8±0.4μs     0.21  bench_core.PackBits.time_packbits_axis1(<class 'bool'>)
```

### Reference:

1. [numpy.packbits usage in python project](https://www.programcreek.com/python/example/102175/numpy.packbits)
2. [Improving performance on comparison algorithm np.packbits](https://stackoverflow.com/questions/48368855/improving-performance-on-comparison-algorithm-np-packbitsa-a-none-axis-1)