@charris 2020-06-02
``` 
/usr/lib/gcc/aarch64-linux-gnu/5/include/arm_neon.h:22314: multiple definition of `_mm_movemask_epi8_neon'
```