@mattip 2020-07-03
I am confused about the connection to #16737. The CI failure here is a segfault on PyPy linux64, not connected to aarch64.

In any case, this looks good to me. Could you rerun the benchmarks on the final code?