@seiko2plus 2020-07-03
@mattip, this crash happens when PyPy running on a machine that comes with SKX features, take a look at **CPU features**
reports before the running of tests. I guess it related to the JIT generator.