@Qiyu8 2020-07-03
Lastest benchmark test result:

testcase | master | pack_bit_neon | ratio
-- | -- | -- | --
time_packbits | 22.0±0.03μs | 5.64±0.02μs | 0.26
time_packbits_axis1 | 392±0.7μs | 69.0±0.1μs | 0.18